Fix yaml to JSON parsing of CF templates; minor CF refactoring (#3424)
