+++ localstack/services/cloudformation/cloudformation_api.py
@@ -1,15 +1,16 @@
 import json
 import logging
 import traceback
+import xmltodict
 from flask import Flask, request
 from requests.models import Response
 from localstack.utils.aws import aws_stack
 from localstack.utils.common import (
-    parse_request_data, short_uid, long_uid, clone, select_attributes, timestamp_millis, recurse_object)
-from localstack.utils.cloudformation import template_deployer
+    parse_request_data, short_uid, long_uid, clone, clone_safe, select_attributes,
+    timestamp_millis, recurse_object)
+from localstack.utils.cloudformation import template_deployer, template_preparer
 from localstack.utils.aws.aws_responses import (
     requests_response_xml, requests_to_flask_response, flask_error_response_xml)
-from localstack.services.cloudformation import cloudformation_listener  # , cloudformation_starter
 
 APP_NAME = 'cloudformation_api'
 app = Flask(APP_NAME)
@@ -60,8 +61,8 @@ class Stack(object):
     def __init__(self, metadata=None, template={}):
         self.metadata = metadata or {}
         self.template = template or {}
-        self._template_raw = clone(self.template)
-        self.template_original = clone(self.template)
+        self._template_raw = clone_safe(self.template)
+        self.template_original = clone_safe(self.template)
         # initialize resources
         for resource_id, resource in self.template_resources.items():
             resource['LogicalResourceId'] = self.template_original['Resources'][resource_id]['LogicalResourceId'] = (
@@ -285,8 +286,8 @@ class StackChangeSet(Stack):
 
 def create_stack(req_params):
     state = RegionState.get()
-    cloudformation_listener.prepare_template_body(req_params)
-    template = template_deployer.parse_template(req_params['TemplateBody'])
+    template_deployer.prepare_template_body(req_params)
+    template = template_preparer.parse_template(req_params['TemplateBody'])
     template['StackName'] = req_params.get('StackName')
     stack = Stack(req_params, template)
     state.stacks[stack.stack_id] = stack
@@ -319,8 +320,8 @@ def update_stack(req_params):
     if not stack:
         return error_response('Unable to update non-existing stack "%s"' % stack_name,
             code=404, code_string='ValidationError')
-    cloudformation_listener.prepare_template_body(req_params)
-    template = template_deployer.parse_template(req_params['TemplateBody'])
+    template_preparer.prepare_template_body(req_params)
+    template = template_preparer.parse_template(req_params['TemplateBody'])
     new_stack = Stack(req_params, template)
     deployer = template_deployer.TemplateDeployer(stack)
     try:
@@ -392,8 +393,8 @@ def list_stack_resources(req_params):
 
 def create_change_set(req_params):
     stack_name = req_params.get('StackName')
-    cloudformation_listener.prepare_template_body(req_params)
-    template = template_deployer.parse_template(req_params['TemplateBody'])
+    template_deployer.prepare_template_body(req_params)
+    template = template_preparer.parse_template(req_params['TemplateBody'])
     template['StackName'] = stack_name
     template['ChangeSetName'] = req_params.get('ChangeSetName')
     stack = existing = find_stack(stack_name)
@@ -452,8 +453,13 @@ def list_imports(req_params):
 
 
 def validate_template(req_params):
-    result = cloudformation_listener.validate_template(req_params)
-    return result
+    try:
+        result = template_preparer.validate_template(req_params)
+        result = '<tmp>%s</tmp>' % result
+        result = xmltodict.parse(result)['tmp']
+        return result
+    except Exception as err:
+        return error_response('Template Validation Error: %s' % err)
 
 
 def describe_stack_events(req_params):
@@ -495,8 +501,8 @@ def get_template_summary(req_params):
         stack = find_stack(stack_name)
         if not stack:
             return stack_not_found_error(stack_name)
-    cloudformation_listener.prepare_template_body(req_params)
-    template = template_deployer.parse_template(req_params['TemplateBody'])
+    template_deployer.prepare_template_body(req_params)
+    template = template_preparer.parse_template(req_params['TemplateBody'])
     req_params['StackName'] = 'tmp-stack'
     stack = Stack(req_params, template)
     result = stack.describe_details()
